requirements.txt
Django==4.2.10
django-admin-charts==1.3.0
django-bower==5.2.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-redis==5.4.0
djangorestframework==3.14.0
pillow==10.2.0
pymemcache==4.0.0
django-cors-headers==4.3.1
